# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/categories": {
      "get": {
        "description": "Devuelve una lista de todas las categorías asociadas al tenant actual.",
        "operationId": "CategoriesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Categorías recuperadas con éxito."
          },
          "401": {
            "description": "No autorizado. Falta el token JWT o es inválido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todas las categorías",
        "tags": [
          "categories"
        ]
      },
      "post": {
        "description": "Crea una nueva categoría asociada al tenant actual.",
        "operationId": "CategoriesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Categoría creada con éxito."
          },
          "400": {
            "description": "Datos de entrada inválidos."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una nueva categoría",
        "tags": [
          "categories"
        ]
      }
    },
    "/categories/{id}": {
      "patch": {
        "description": "Actualiza los datos de una categoría existente para el tenant actual.",
        "operationId": "CategoriesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCategoryDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Categoría actualizada con éxito."
          },
          "404": {
            "description": "Categoría no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una categoría",
        "tags": [
          "categories"
        ]
      },
      "delete": {
        "description": "Elimina una categoría asociada al tenant actual.",
        "operationId": "CategoriesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Categoría eliminada con éxito."
          },
          "404": {
            "description": "Categoría no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una categoría",
        "tags": [
          "categories"
        ]
      }
    },
...
    "/products": {
      "get": {
        "description": "Devuelve una lista de productos para el tenant actual.",
        "operationId": "ProductsController_getProducts",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Productos recuperados con éxito."
          },
          "401": {
            "description": "No autorizado. Falta el tenantId o el token JWT es inválido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todos los productos",
        "tags": [
          "products"
        ]
      },
      "post": {
        "description": "Crea un producto asociado al tenant actual.",
        "operationId": "ProductsController_createProduct",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Producto creado con éxito."
          },
          "400": {
            "description": "Datos de entrada inválidos."
          },
          "401": {
            "description": "No autorizado. Falta el tenantId o el token JWT es inválido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear un nuevo producto",
        "tags": [
          "products"
        ]
      }
    },
    "/sales": {
      "get": {
        "description": "Devuelve una lista de todas las ventas asociadas al tenant actual.",
        "operationId": "SalesController_getSales",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Ventas recuperadas con éxito."
          },
          "401": {
            "description": "No autorizado. Falta el token JWT o es inválido."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener todas las ventas",
        "tags": [
          "sales"
        ]
      },
      "post": {
        "description": "Crea una nueva venta asociada al tenant actual.",
        "operationId": "SalesController_createSale",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSaleDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venta creada con éxito."
          },
          "400": {
            "description": "Datos de entrada inválidos."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Crear una nueva venta",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/shift/{shiftId}": {
      "get": {
        "description": "Devuelve las ventas asociadas a un turno específico del tenant actual.",
        "operationId": "SalesController_getSalesByShift",
        "parameters": [
          {
            "name": "shiftId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Ventas por turno recuperadas con éxito."
          },
          "404": {
            "description": "Turno no encontrado."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obtener ventas por turno",
        "tags": [
          "sales"
        ]
      }
    },
    "/sales/{tenantId}/{saleId}": {
      "put": {
        "description": "Actualiza los datos de una venta específica para el tenant especificado.",
        "operationId": "SalesController_updateSale",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "saleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSaleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venta actualizada con éxito."
          },
          "404": {
            "description": "Venta no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Actualizar una venta",
        "tags": [
          "sales"
        ]
      },
      "delete": {
        "description": "Elimina una venta asociada al tenant especificado.",
        "operationId": "SalesController_deleteSale",
        "parameters": [
          {
            "name": "tenantId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "saleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venta eliminada con éxito."
          },
          "404": {
            "description": "Venta no encontrada."
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Eliminar una venta",
        "tags": [
          "sales"
        ]
      }
    },
...
  },
  "info": {
    "title": "Inventory POS API",
    "description": "API para el sistema de inventario y punto de venta",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre de la categoría",
            "example": "Electrónica"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateCategoryDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre actualizado de la categoría",
            "example": "Ropa"
          }
        }
      },
...
      "CreateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre del producto",
            "example": "Laptop"
          },
          "price": {
            "type": "number",
            "description": "Precio del producto",
            "example": 999.99
          },
          "categoryId": {
            "type": "string",
            "description": "ID de la categoría del producto",
            "example": "63f1b23c9a7d2e1234abcd56"
          }
        },
        "required": [
          "name",
          "price",
          "categoryId"
        ]
      },
      "SaleItemDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string",
            "description": "ID del producto",
            "example": "63f1b23c9a7d2e1234abcd56"
          },
          "quantity": {
            "type": "number",
            "description": "Cantidad del producto vendido",
            "example": 2
          },
          "price": {
            "type": "number",
            "description": "Precio unitario del producto",
            "example": 15.99
          }
        },
        "required": [
          "productId",
          "quantity",
          "price"
        ]
      },
      "CreateSaleDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Lista de productos vendidos",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total de la venta",
            "example": 31.98
          }
        },
        "required": [
          "items",
          "total"
        ]
      },
      "UpdateSaleDto": {
        "type": "object",
        "properties": {
          "items": {
            "description": "Lista actualizada de productos vendidos",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SaleItemDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total actualizado de la venta",
            "example": 31.98
          }
        }
      },
...
    }
  }
}